+++ client-sdks/client-js/src/route-types.generated.ts
@@ -301,11 +301,7 @@ type Shared_Type_16 = {
   };
   /** Which processor phases to enable */
   enabledPhases: (
-    | 'processInput'
-    | 'processInputStep'
-    | 'processOutputStream'
-    | 'processOutputResult'
-    | 'processOutputStep'
+    'processInput' | 'processInputStep' | 'processOutputStream' | 'processOutputResult' | 'processOutputStep'
   )[];
 };
 
@@ -16554,11 +16550,7 @@ export type GetProcessorProviders_Response = {
     name: string;
     description?: string | undefined;
     availablePhases: (
-      | 'processInput'
-      | 'processInputStep'
-      | 'processOutputStream'
-      | 'processOutputResult'
-      | 'processOutputStep'
+      'processInput' | 'processInputStep' | 'processOutputStream' | 'processOutputResult' | 'processOutputStep'
     )[];
   }[];
 };
@@ -16591,11 +16583,7 @@ export type GetProcessorProvidersProviderId_Response = {
   name: string;
   description?: string | undefined;
   availablePhases: (
-    | 'processInput'
-    | 'processInputStep'
-    | 'processOutputStream'
-    | 'processOutputResult'
-    | 'processOutputStep'
+    'processInput' | 'processInputStep' | 'processOutputStream' | 'processOutputResult' | 'processOutputStep'
   )[];
   configSchema: {
     [key: string]: unknown;